pyproject.toml
[project]
name = "efb-wechat-mprss"
version = "0.1.0"
description = "Generate RSS feeds for WeChat public accounts from EFB message logs"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pyyaml>=6.0",
]

[project.scripts]
mprss = "backend.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "frontend"]
